Leverage url param to avoid copy/pasted html files
Files in the "plan" directory where all copy/pasted. This makes it tedious
for the maintenance.
For instance because those html don't have the
same html input fields as new.html and hence the JS needs to take into
account different cases (corollary: after that commit, the JS can be
simplified in a subsequent commit).
Other example: if at some point we want to avoid using files from CDN
(in order to improve the GDPR compliance) then we would have needed to
update each of those similar html files.

This commit leverages the ability to load a file from a path specified in
the URL in order to avoid having to duplicate all those files.